Fix #737: Remove ImageInfo() LongInt overload, use Integer version only
This partially reverts 6c13fc0c. It's the same problem as with
ScreenInfo() (57335003) and GetJoyStick() (87b1c3fe): In a call such as
	ImageInfo image, , , , , pixdata
the compiler cannot decide whether to use the Long or LongInt overload
because none of the integer parameters are given.

To allow such calls to keep working, there can only be one version of
ImageInfo(). As with the other cases it seems best to choose a version
with Integer parameters. Besides being FB's default data type, it will
accept Longs/Integers on 32bit (as always) and LongInts/Integers on 64bit.
That's not perfect as existing code using Longs on 32bit won't compile for
64bit, it must be changed to use Integers first, but this is acceptable.

+++ src/gfxlib2/gfx_image_info.c
@@ -2,10 +2,16 @@
 
 #include "fb_gfx.h"
 
-FBCALL int fb_GfxImageInfo( void *img,
-							int *width,	int *height,
-							int *bpp,	int *pitch,	void **imgdata,
-							int *size)
+FBCALL int fb_GfxImageInfo
+	(
+		void *img,
+		ssize_t *width,
+		ssize_t *height,
+		ssize_t *bpp,
+		ssize_t *pitch,
+		void **imgdata,
+		ssize_t *size
+	)
 {
 	PUT_HEADER *header;
 	int bpp_, width_, height_, pitch_, headerSize_;
